[sysvabi64] Document requirements for tools wrt BTI (#282)
Add requirements for when a tool must generate a BTI instruction. This
permits tools to elide BTI instructions when they can prove that no
indirect branch to that location is possible from local information
available to the tool.

Static linkers are not allowed to assume that all direct branch targets
have a BTI instruction. If a veneer is required then the static linker
must generate additional BTI instructions if needed.

A static linker is allowed to assume that a symbol that is exported to
the dynamic symbol table has a BTI instruction.

In practice this will permit compilers to remove BTI instructions from
static functions that do not have their address taken and that address
escapes the function.

This matches the behavior of the GNU toolchain.

Fixes https://github.com/ARM-software/abi-aa/issues/196