[pauthabi64][aaelf64] Move PAuth GOT relocs out of private space
In the initial draft relocations for signing GOT entries were
defined in the private space reserved for experiments. It was not
known if these relocations would be useful as if RELRO is supported
the GOT does not need to be signed. There is now at least one user
of the GOT signing relocations so we should move the relocations
out of the private space and into the space reserved for the
PAuthABI.

This PR includes relocations added in
https://github.com/ARM-software/abi-aa/pull/295

Fixes: https://github.com/ARM-software/abi-aa/issues/298