[buildattr64] Recommend aeabi_pauthabi is only emitted when needed
The aeabi_pauthabi subsection is mandatory, which means that if it
is present an object consumer is required to understand the contents
of the subsection. That makes a relocatable object containing
an aeabi_pauthabi subsection with explicit 0 values for the tags
less portable than relying on the implicit 0 values when the
subsection isn't present.