Markup: fix two accidental triple-backticks. (#303)
I noticed one of these in passing in RTABI32: Github's Markdown renderer
showed a spurious backtick at the end of a sensible identifier in
fixed-pitch text. Looking at the source, that's because the fixed-pitch
section was closed by three backticks instead of the usual two.

Grepping the rest of the repository turned up one more case of the same
thing, in DBGOVL32. This patch fixes both.