[pauthabi] Provide decimal representation of relocation code
Add a decimal representation of the relocation code alongside the
relocation code for easier cross-reference with AAELF64.

The PAuthABI uses hex for the relocation codes as the only
implementation LLVM uses hex in the relocation definitions. This
makes it easier to cross-reference. Adding the decimal in
parentheses satisfies both forms of cross-reference.