[AAPCS64] Clarify how __bf16 affects HFAs
Currently, it's not completely clear whether a struct containing a
mixture of different 16-bit floating point types is a Homogeneous
Aggregate or not:

  struct foo {
    __fp16 a;
    __bf16 b;
  };

I think that the best interpretation of the current text is that this is
an HFA, because there is only one entry in the "Fundamental Data Types"
table, which is the only thing considered in the definition of a
Homogeneous Aggregate.

However, the table in the "Mapping of C & C++ built-in data types"
section used the formats in a column which otherwise contains
Fundamental Data Types, which could be read to imply that the above
struct is _not_ an HFA. I don't think this is correct, so I've moved the
formats to the notes column, making it clear that the two source types
map to the same Fundamental Data Type.

Current compiler behaviour:
* Clang considers this struct to be an HFA, matches the behaviour
  described in this ABI patch.
* GCC does not consider it to be an HFA, but it also does not think that
  a struct only containing __bf16 is an HFA either. That's unambiguously
  incorrect both before and after this patch, so it's just a special
  case of a broader GCC bug.