[buildattr64] Reword PAuthABI build attributes.
Make it clear that the platform and schema attributes must be
combined as a tuple.

Make (platform = 0, schema = 1) the reserved value guaranteed to
be incompatible with any other use of the PAuthABI.

Make the default values (platform = 0, schema = 0) represent no
information, distinct from the reserved "incompatible" tuple.

Permit schema = 0 to be a valid identifier for platform N, where N
is non-zero.

Clarify that combining (0, 0) with (N, M) where N is non-zero is
implementation defined. This permits something similar to
-zbtireport=(none, warn, error).

Clarify mappings to and from the .note.gnu.property section.