[sysvabi64] document TLSDESC resolver extension register reqs
Document the decision in the GCC mailing list thread
TLSDESC clobber ABI stability/futureproofness?
https://gcc.gnu.org/legacy-ml/gcc/2018-10/msg00112.html

TLSDESC resolver functions assume that any registers added
by an extension are caller saved for a TLSDESC call.

A brief summary:

Dynamic TLS may be lazy allocated upon the first use of a TLSDESC
resolver. This may involve calls to heap allocation functions
provided by the user, which may use registers from extensions
like SVE and SME. As the resolver function can't know what is
saved it would have to save all SVE and SME state. This would
be way more expensive than a caller save, and an older libc
written prior to the introduction of the extension would be
unaware of them so the caller has to do the save.

* The SVE and SME state is already