[AAELF64] BTI and veneers for bare-metal contexts.
Range extension veneers in a bare-metal context must not assume a
BTI instruction is at the destination.

There is documentation in sysvabi64 that states this, but strictly
speaking this need not apply to a bare-metal context. Document that
linker generated veneers cannot assume a BTI for definitions in the
same link unit.

Bare-metal systems do not support PLT or iPLT (ifunc) so the
requirements surrounding code-generation can stay in the sysvabi64.