[SYSVABI] Add example that ifunc resolvers are indirectly called
An ifunc resolver function needs a BTI as it is called indirectly
via the dynamic loader when resolving the R_AARCH64_IRELATIVE
relocation. When static linking the relocation is also resolved
via an indirect branch to the resolver from a local resolver
function like _dl_relocate_static_pie in glibc.