[sysvabi64] Require PT_GNU_PROPERTY when program properties used. (#301)
Added a requirement that a static linker must generate a PT_GNU_PROPERTY
program header for an executable or shared-library that contains program
properties.

The linux kernel and glibc are currently relying on the program header
to find the program properties and are not attempting to search through
the consolidated notes section (PT_NOTE) program header if it isn't
present.

fixes: https://github.com/ARM-software/abi-aa/issues/299