[AAELF64] Simplify expressions used for relative relocations
The function Delta(S) returns the current binary's load bias (for the
null symbol as well as for locally-defined S), and is only used in the
definition of RELATIVE family relocations, which are not expected to
have a non-null symbol operand.

Theoretically, a non-null symbol operand could be used to refer to a
symbol in another binary, which would then cause Delta(S) to evaluate
to the load bias of that binary. But I am unaware of any dynamic loader
implementing this (checked glibc/musl/bionic/FreeBSD/NetBSD/OpenBSD),
and its utility seems limited.

The description of R_<CLS>_RELATIVE appears to contemplate a different
case in which different segments in the same binary have different load
biases. It is unclear to me how this would work in practice. If the
idea is to accommodate individual segments being loaded at independent
addresses, I don't think this would be enough; ELF is not generally
designed to accommodate this so numerous other changes to relocation
processing would need to be made in order for this to work.

Therefore, simplify the definition of Delta as well as the users, and
bring it in line with existing practice, by removing the argument and
having it always produce the current binary's load bias, and adjust the
R_<CLS>_RELATIVE description to match.