Changes from last revision:
Remove the mention of SME support routines and __builtin_cpu_supports
as there is no obligation that they use __aarch64_cpu_features. Instead
say that the variable can only be placed in RELRO if it is only accessed
by FMV resolvers. Otherwise, if used by compiler generated code or the
runtime library then it requires constructor init and cannot be in RELRO.