[aadwarf64]] Add DW_CFA_AARCH64_set_ra_state stack frame operation
The DW_CFA_AARCH64_set_ra_state operation updates the RA_SIGN_STATE
pseudo register with the current signing state. If the signing
state includes signing is DW_AARCH64_RA_SIGNED_SP_PC, then it also
provides the offset to the signing instruction so that the PC value
used in the signing can be calculated.

The DW_CFA_AARCH64_negate_ra_state_with_pc operation has
been marked as deprecated.  This is because it has been
found that it is not suitable for describing all cases
where the PC was used to sign the return address (see
https://github.com/ARM-software/abi-aa/issues/327)

The contents of the RA_SIGN_STATE pseudo register is also changed
from being described in terms of a set of bits to being a series
of defined values.

Previously the state of the RA_SIGN_STATE pseudo register was
changed implicitly by the DW_CFA_AARCH64_negate_ra_state and
DW_CFA_AARCH64_set_ra_state operations. This meant that the actual
encoding was actually internal to any implementation.

Now with the introduction of the DW_CFA_AARCH64_set_ra_state
operation the encoding has been made externally visible.  So the
opportunity has been taken now to change the encoding to a simpler
form.