[AAELF64] Document structure protection relocations (#340)
The structure protection extension [1], introduces two new relocations:

R_AARCH64_PATCHINST which replaces an instruction with the result of the
relocation if the referenced symbol is defined. The relocation is
intended to implement deactivation symbols [2], which can be used to
"undo" structure protection when it isn't safe to apply.

R_AARCH64_FUNCINIT64 generates a R_AARCH64_IRELATIVE relocation at the
place. This is similar to R_AARCH64_ABS64 referencing a symbol of
binding STB_GNU_UNIQUE, with the main difference being that
R_AARCH64_FUNCINIT64 references non STB_GNU_UNIQUE symbols. The
relocation is used to initialize function pointers in static structs at
load time by calling the equivalent of an ifunc resolver function that
returns the signed address.

[1]
https://discourse.llvm.org/t/rfc-structure-protection-a-family-of-uaf-mitigation-techniques/85555
[2] https://discourse.llvm.org/t/rfc-deactivation-symbols/85556