[sysvabi64] Code Models. Limit range to GOT to 2GiB
The R_AARCH64_GOTPCREL32 relocation, currently used in clang when
relative vtables are enabled, has a signed offset so is limited to
2 GiB range. This limits the range of GOT from the vtables to 2GiB.

Document this restriction in the tables for the small and medium
code model. Add a note to explain the reason, so that if code
generation does not use the relocation the distance to the GOT can
be increased.

Fixes #343