[sysvabi64] Move TLSDESC resolvers to separate design doc
The TLSDESC resolver functions are not ABI so we can move them
out of the sysvabi64 document. Providing some examples that can
be used by a dynamic linker is still useful so move this to the
design documents section.

Add a comment about DTV surplus TLS that permits a dynamic loader
to dlopen a DSO with initial-exec TLS. There can be a small
number of performance critical shared-libraries that use initial
exec TLS, but are expected to be opened via dlopen, particularly
by scripting languages like python.