[FMV][AArch64] Document the interface for Function Multi-versioning. (#332)
Documents the Application Binary Interface requirements for
Function Multi-versioning (FMV) on System V platforms.

A global variable `__aarch64_cpu_features` provided by
the runtime library contains information about the available
CPU features. The variable must be defined as DSO-local with
its symbol visibility set to `STV_HIDDEN`.

A function `__init_cpu_features_resolver` provided by
the runtime library is used for initializing the variable
`__aarch64_cpu_features`. The function must be defined as
DSO-local with its symbol visibility set to `STV_HIDDEN`.

Accessing `__aarch64_cpu_features` is reserved for the
compiler generated code or the runtime library. If the
variable is only accessed for FMV, then it may be placed
in the Relocation Read Only (RELRO) program segment to
prevent it from being modified. Otherwise, if the variable
is used besides FMV the runtime library must ensure that
it is initialized via a constructor function and cannot
be placed in the `RELRO` segment.